+++ MWPushNotificationsPlugin/MWPushNotificationsPlugin/Steps/Push Notifications/MWPushNotificationsViewController.swift
@@ -10,18 +10,6 @@ import UserNotifications
 import Combine
 import MobileWorkflowCore
 
-enum L10n {
-    enum PushNotification {
-        static let generalError = "Unable to register for Push Notifcations."
-        static let deniedTitle = "Push Notifications"
-        static let deniedText = "You have currently opted out of receiving Push Notifications. You can update this preference in Settings."
-        static let deniedCancelTitle = "Skip"
-        static let deniedConfirmTitle = "Settings"
-        static let enableButtonTitle = "Enable"
-        static let skipButtonTitle = "Skip"
-    }
-}
-
 enum MWPushNotificationsError: LocalizedError {
     case registrationTimeout
